refactor(core): Move models used in clouddriver-web to clouddriver-core
