+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/edda/EddaApi.groovy
@@ -18,7 +18,7 @@ package com.netflix.spinnaker.clouddriver.aws.edda
 
 import com.amazonaws.services.elasticloadbalancingv2.model.Listener
 import com.netflix.spinnaker.clouddriver.aws.model.edda.EddaRule
-import com.netflix.spinnaker.clouddriver.aws.model.edda.LoadBalancerInstanceState
+import com.netflix.spinnaker.clouddriver.model.edda.LoadBalancerInstanceState
 import com.netflix.spinnaker.clouddriver.aws.model.edda.TargetGroupAttributes
 import com.netflix.spinnaker.clouddriver.aws.model.edda.TargetGroupHealth
 import retrofit.http.GET
