+++ clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/model/InstanceTargetGroups.groovy
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.aws.model
+package com.netflix.spinnaker.clouddriver.model
 
-import com.netflix.spinnaker.clouddriver.model.HealthState
 import groovy.transform.CompileStatic
 import groovy.transform.EqualsAndHashCode
 import groovy.transform.Immutable
-import groovy.transform.TypeCheckingMode
 
 @CompileStatic
 @Immutable
