+++ clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/model/edda/LoadBalancerInstance.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Lookout, Inc.
+ * Copyright 2014 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.aws.model;
+package com.netflix.spinnaker.clouddriver.model.edda
 
-import java.util.Collection;
+import groovy.transform.Canonical
 
-public interface Role {
-
-  String getName();
-  String getId();
-  Collection<? extends TrustRelationship> getTrustRelationships();
+@Canonical
+class LoadBalancerInstance {
+  String instanceId
+  String state
+  String reasonCode
+  String description
 }
