+++ clouddriver-core/src/test/groovy/com/netflix/spinnaker/clouddriver/model/edda/InstanceLoadBalancerStateSpec.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.aws.model.edda
+package com.netflix.spinnaker.clouddriver.model.edda
 
 import spock.lang.Specification
